DONE: We need to compare how the values in the dependencyValues vetor would have evolved if we were deleting the modNodes instead of the subMod nodes. Write a separate method to perform this test.
NEXT: Also, we need to write another method that removes nodes only if they are not present in the envelop of 100 nodes as well.

NEXT: Find out how to visualize data. If a nodes i being removed, then how far is it from the seed Set. Check SNAP as well. See if you can figure something out from there.